% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ag_hunt_nature.R
\name{ag_hunt_nature}
\alias{ag_hunt_nature}
\title{Agricultural, hunting, forestry, fishing indicator}
\arguments{
\item{data}{the economicIndicators data set}

\item{country}{the country to input}

\item{year}{the year set for the country}
}
\description{
Determines  the indicator for Agricultural, hunting, forestry, fishing
}
\details{
Takes in the data and selects a country set a certain year to find that indicator
in that country in a certain year.
}
\examples{
ag_hunt_nature(economicIndicators_data,'Spain', 1970)
}
\keyword{internal}
