% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ama_exchange_rate.R
\name{ama_exchange_rate}
\alias{ama_exchange_rate}
\title{Country AMA exchange rate}
\arguments{
\item{data}{the economicIndicators data set}

\item{country}{the country for the specific AMA exchange rate}

\item{year}{the specific year for the country}
}
\description{
Provides a country's AMA exchange rate in a specific year
}
\details{
Takes in the data and has to check the range of the year before selecting the data and providing
the GDP per population in a country.
}
\examples{
ama_exchange_rate(economicIndicators_data,'Spain', 1970)
}
