\name{country_GDP_per_pop_timed}
\alias{country_GDP_per_pop_timed}
\title{Country GDP Per Population over Time}
\usage{
country_GDP_per_pop_timed()
}
\description{
Prints out a linear model of a country's GDP per population over time.
}
\examples{
country_GDP_per_pop_timed(economicIndicators_data,'Spain')
}
