% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/economicIndicators_data.R
\docType{data}
\name{economicIndicators_data}
\alias{economicIndicators_data}
\title{Economic indicators dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10512 rows and 26 columns.
}
\usage{
economicIndicators_data
}
\description{
Determines a country's GDP per population over time.
}
\details{
Data that shows different economic indicators such as population, GDP, and the ama exchange rate.
}
\keyword{datasets}
