% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdp_per_pop1yr.R
\name{gdp_per_pop1yr}
\alias{gdp_per_pop1yr}
\title{Population country Comparison}
\arguments{
\item{data}{the economicIndicators data set}

\item{country}{the country for the specific GDP per population}

\item{year}{the specific year for the country}
}
\description{
Compares a country's population in a specific year.
}
\details{
Takes in the data and has to check the range of the year before selecting the data and providing
the GDP per population in a country.
}
\examples{
gdp_per_pop1yr(economicIndicators_data,'Spain', 1970)
}
