% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_gdp_per_year.R
\name{highest_gdp_per_year}
\alias{highest_gdp_per_year}
\title{Highest GDP value based on the Given year}
<<<<<<< HEAD
\usage{
highest_gdp_per_year(data, year)
}
=======
>>>>>>> 1a5565eab545dadbb3f8e44baf45497d1eeebb7d
\arguments{
\item{data}{the economicIndicators data set}

\item{year}{the year set for the country}
}
\description{
Determines the country with the highest GDP based on the given year
}
\examples{
highest_gdp_per_year(economicIndicators_data, 1970)
}
