% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowest_gdp_per_year.R
\name{lowest_gdp_per_year}
\alias{lowest_gdp_per_year}
\title{Lowest GDP value based on the Given year}
\arguments{
\item{data}{the economicIndicators data set}

\item{year}{the year set for the country}
}
\description{
Determines the country with the lowest GDP based on the given year
}
\details{
Filters a year and then returns the top one using head(1)
}
\examples{
lowest_gdp_per_year(economicIndicators_data, 1970)
}
\keyword{internal}
