% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_gdp_per_year.R
\name{mean_gdp_per_year}
\alias{mean_gdp_per_year}
\title{Mean GDP value based on the Given year}
\arguments{
\item{data}{the economicIndicators data set}

\item{year}{the year set for the country}
}
\description{
Determines mean GDP of the given year
}
\details{
Filters a year and then finds the mean of of the GDP collumn
}
\examples{
mean_gdp_per_year(economicIndicators_data, 1970)
}
\keyword{internal}
