% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_country_compare_1year.R
\name{pop_country_compare_1year}
\alias{pop_country_compare_1year}
\title{Population country Comparison in one year}
\arguments{
\item{data}{the economicIndicators data set}

\item{country1}{the first country for comparison}

\item{country2}{the second country for comparison}

\item{year}{the year set for comparison of the countries}
}
\description{
Compares a country's population in a specific year.
}
\details{
Takes in the data and has to check the range of the year before selecting the data and providing
the population in both countries.
}
\examples{
pop_country_compare_1year(economicIndicators_data,'Spain', 'Italy', 1970)
}
\keyword{internal}
