% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_GDP_per_pop_timed.R
\name{country_GDP_per_pop_timed}
\alias{country_GDP_per_pop_timed}
\title{Country GDP per population over Time}
\arguments{
\item{data}{economicIndicators dataset}

\item{country}{the country for the specific GDP per population}
}
\description{
Determines a country's GDP per population over time.
}
\details{
Computes the specific country's GDP per population by dividing the GDP by the population
and generates  with a timed series graph to show the change. The statistics provides information
about the GDP per population association with each year.
}
\examples{
country_GDP_per_pop_timed(economicIndicators_data,'Spain')
}
\keyword{internal}
